+++ src/main/java/com/sparta/poly/PolyInterface.java
@@ -1,5 +1,5 @@
 package com.sparta.poly;
 
 public interface PolyInterface {
-    public int getArea();
+    public double getArea();
 }
