+++ src/main/java/com/sparta/poly/Square.java
@@ -1,13 +1,11 @@
 package com.sparta.poly;
 
 public class Square extends Polygon{
-
     Square(int sideL) {
         super("Square", 4, sideL);
     }
     @Override
-    public int getArea() {
+    public double getArea() {
         return sideLength*sideLength ;
     }
-
 }
